Upd: refactoring, logic
